+++ fedbiomed/researcher/exceptions.py
@@ -1,6 +1,0 @@
-class DefaultStrategyException(Exception):
-    pass
-
-
-class TrainingException(Exception):
-    pass